Refactoring components to make them reusable
