--- components/blogPosts/blogPostsHead.js
@@ -1,12 +1,0 @@
-import Head from "next/head";
-
-
-export function BlogPostsHead() {
-  return (
-    <>
-        <Head>
-            <title>{"//"}Blog Posts</title>
-        </Head>
-    </>
-  );
-}
