--- components/contactDetails/contactDetailsHead.js
@@ -1,11 +1,0 @@
-import Head from "next/head";
-
-export function ContactDetailsHead() {
-  return (
-    <>
-      <Head>
-        <title>{"//"}Contact Details</title>
-      </Head>
-    </>
-  );
-}
