--- components/index/indexH2.styled.jsx
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const IndexH2 = styled.h2`
-font-size: 24px;
-font-weight: bold;
-color: #333;
-margin-bottom: 30px;
-`
