--- components/index/indexHead.js
@@ -1,12 +1,0 @@
-import Head from "next/head";
-
-export function IndexHead() {
-  return (
-    <>
-      <Head>
-        <title key="title">{"//"}Home</title>
-        <meta property="og:image" content="https://res.cloudinary.com/dhfmjugt0/image/upload/v1680130330/Screenshot_from_2023_03_30_09_21_52_0de5e91286.png?updated_at=2023-03-29T22:52:11.653Z" />
-      </Head>
-    </>
-  );
-}
