--- components/index/indexItem.styled.jsx
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-import { motion } from "framer-motion";
-
-export const IndexItem = styled(motion.li)`
-margin-bottom: 0px;
-`
