--- components/index/indexList.styled.jsx
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const IndexList = styled.ul`
-list-style: none;
-padding: 0;
-margin: 0;
-`
