--- components/index/indexPostTitle.styled.jsx
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const IndexPostTitle = styled.h3`
-margin-top: 0;
-font-size: 24px;
-font-weight: bold;
-color: #333;
-margin-bottom: 5px;
-`
