--- components/index/indexSubtitle.styled.jsx
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const IndexSubtitle = styled.p`
-font-size: 16px;
-color: #666;
-margin-bottom: 0;
-}
-`
