--- components/index/latestBlog.styled.jsx
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export const LatestBlog = styled.div`
-  padding-top: 50px;
-  padding-bottom: 25px;
-`
