--- components/index/outerContainer.styled.jsx
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const OuterContainer = styled.div`
-  max-width: 1500px;
-  margin: 0 auto;
-  padding: 50px;
-  display: flex;
-  flex-direction: column;
-`
