+++ components/pages/head.js
@@ -1,10 +1,11 @@
 import Head from "next/head";
 
-export function CustomHead({ text }) {
+export function CustomHead({ text, meta }) {
   return (
     <>
       <Head>
         <title>{text}</title>
+        {meta && <meta property={meta.property} content={meta.content} />}
       </Head>
     </>
   );
