--- components/portfolioGame/portfolioGameHead.js
@@ -1,11 +1,0 @@
-import Head from 'next/head';
-
-export function PortfolioGameHead() {
-    return (
-        <>
-            <Head>
-                <title key="title">{"//"}Portfolio Game</title>
-            </Head>
-        </>
-    )
-}
