+++ pages/portfoliogame.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { PortfolioGameBody } from '@/components/portfolioGame/portfolioGameBody.js';
-import { PortfolioGameHead } from '@/components/portfolioGame/portfolioGameHead.js';
+import { CustomBody } from '@/components/pages/body.js';
+import { CustomHead } from '@/components/pages/head.js';
 import { PortfolioGameCode } from '@/components/portfolioGame/portfolioGameCode.js';
 import { PortfolioGameIframe } from '@/components/portfolioGame/portfolioGameIframe';
 import { PortfolioGameTitle } from '@/components/portfolioGame/portfolioGametitle';
@@ -35,15 +35,22 @@ box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
 transition: box-shadow 0.3s ease;
 `
 
+const body = "I created a portfolio game using Python that I\'m really proud of. The game is simple and consists of three unique rooms that showcase parts of my life experiences. It serves as an interactive resume.\n\n While working on this project, I gained a lot of experience in Python programming, including data structures, functions, and game design.\n I also learned about game development concepts such as collision detection, game logic, and game state management.\n\n The game has been a valuable addition to my portfolio, highlighting my programming skills and creativity. I wrote an article about the development process, which explains the technical details and design choices I made. If you\'re interested in game development or Python programming, feel free to check it out"
+
 export default function PortFoliogame() {
   return (
     <Layout>
       <Container>
         <Article>
-        <PortfolioGameHead />
+        <CustomHead text="//Portfolio Game"/>
         <PortfolioGameTitle />
         <PortfolioGameIframe />
-        <PortfolioGameBody />
+        <CustomBody
+          text={body}
+          linkHref="/python-resume-game"
+          linkText="here!"
+        />        
+
         <PortfolioGameCode />
         </Article>
       </Container>
